app/requirements.txt
tensorflow==2.17.1
pillow
fastapi
uvicorn
